refactor: Move WellSchedule & related methods to new file
