+++ completor/create_wells.py
@@ -6,7 +6,7 @@ import numpy as np
 import numpy.typing as npt
 import pandas as pd
 
-from completor import completion
+from completor import completion, wells2
 from completor.constants import Headers, Method
 from completor.logger import logger
 from completor.read_casefile import ReadCasefile
@@ -52,7 +52,7 @@ class CreateWells:
         self.case: ReadCasefile = case
         self.method = _get_method(self.case)
 
-    def update(self, well_name: str, schedule: completion.WellSchedule) -> None:
+    def update(self, well_name: str, schedule: wells2.WellSchedule) -> None:
         """Update class variables in CreateWells.
 
         Args:
@@ -65,7 +65,7 @@ class CreateWells:
         active_laterals = _get_active_laterals(well_name, self.case.completion_table)
         for lateral in active_laterals:
             self.df_completion = self.case.get_completion(well_name, lateral)
-            self.df_welsegs_header, self.df_welsegs_content = completion.get_well_segments(
+            self.df_welsegs_header, self.df_welsegs_content = wells2.get_well_segments(
                 schedule.msws, well_name, lateral
             )
 
@@ -186,7 +186,7 @@ def _get_active_laterals(well_name: str, df_completion: pd.DataFrame) -> npt.NDA
     return np.array(df_completion[df_completion[Headers.WELL] == well_name][Headers.BRANCH].unique())
 
 
-def _select_well(well_name: str, schedule: completion.WellSchedule, lateral: int) -> pd.DataFrame:
+def _select_well(well_name: str, schedule: wells2.WellSchedule, lateral: int) -> pd.DataFrame:
     """Filter the reservoir data for this well and its laterals.
 
     Args:
@@ -197,8 +197,8 @@ def _select_well(well_name: str, schedule: completion.WellSchedule, lateral: int
     Returns:
         Filtered reservoir data.
     """
-    df_compsegs = completion.get_completion_segments(schedule.msws, well_name, lateral)
-    df_compdat = completion.get_completion_data(schedule.msws, well_name)
+    df_compsegs = wells2.get_completion_segments(schedule.msws, well_name, lateral)
+    df_compdat = wells2.get_completion_data(schedule.msws, well_name)
     df_reservoir = pd.merge(df_compsegs, df_compdat, how="inner", on=[Headers.I, Headers.J, Headers.K])
 
     # Remove WELL column in the df_reservoir.
