+++ tests/completor/test_completion.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from completor import completion  # type: ignore
+from completor import completion, wells2
 from completor.constants import Headers, Keywords, Method
 from completor.exceptions import CompletorError
 
@@ -979,17 +979,17 @@ def test_skin():
             Headers.RO,
         ],
     )
-    well_schedule = completion.WellSchedule(np.array(["A1"]))
-    well_schedule.msws = completion.handle_compdat(well_schedule.msws, set(well_schedule.active_wells), compdat)
+    well_schedule = wells2.WellSchedule(np.array(["A1"]))
+    well_schedule.msws = wells2.handle_compdat(well_schedule.msws, set(well_schedule.active_wells), compdat)
     df_out = well_schedule.msws["A1"][Keywords.COMPDAT]
     pd.testing.assert_frame_equal(df_out, df_true)
 
 
 def test_set_welsegs_negative_length_segments(caplog):
     """Test that negative segments inside a branch give a warning."""
-    schedule = completion.WellSchedule(np.array(["A1"]))
+    schedule = wells2.WellSchedule(np.array(["A1"]))
 
-    schedule.msws = completion.set_welsegs(
+    schedule.msws = wells2.set_welsegs(
         schedule.msws,
         schedule.active_wells,
         [
